moved logic to Invocation and checking also non-mocked methods
